+++ src/curvesimilarities/integfrechet.py
@@ -4,6 +4,7 @@ import numpy as np
 from numba import njit
 
 from ._algorithms.ifd import _ifd_acm, _ifd_acm_1d, _ifd_owp
+from .util import index2arclength
 
 __all__ = [
     "ifd",
@@ -101,7 +102,7 @@ def ifd(P, Q, delta, dist="euclidean"):
 
 
 @njit(cache=True)
-def ifd_owp(P, Q, delta, dist="euclidean"):
+def ifd_owp(P, Q, delta, dist="euclidean", param="arc-length"):
     """Integral Fréchet distance and its optimal warping path.
 
     Parameters
@@ -116,6 +117,8 @@ def ifd_owp(P, Q, delta, dist="euclidean"):
         Maximum length of edges between Steiner points. Refer to :func:`ifd`.
     dist : {'euclidean', 'squared_euclidean'}
         Type of :math:`dist`. Refer to :func:`ifd`.
+    param : {'arc-length', 'vertex'}
+        Parametrization of matching.
 
     Returns
     -------
@@ -143,5 +146,18 @@ def ifd_owp(P, Q, delta, dist="euclidean"):
         ifd = NAN
     else:
         ifd = L[-1, -1]
-    path = _ifd_owp(P, Q, B, L, delta, dist)
-    return ifd, path[::-1]
+    path = _ifd_owp(P, Q, B, L, delta, dist)[::-1]
+
+    if param == "arc-length":
+        path = np.stack(
+            (
+                index2arclength(P, path[:, 0].copy()),
+                index2arclength(Q, path[:, 1].copy()),
+            )
+        ).T
+    elif param == "vertex":
+        pass
+    else:
+        raise ValueError("Unknown option for parametrization.")
+
+    return ifd, path